Move some components around
