+++ src/components/upload/FileDropper.tsx
@@ -1,8 +1,8 @@
-import useError from "../../../hooks/useError";
-import useUpload from "../../../hooks/useUpload";
-import { looksLikeIsi } from "../../../utils/isi";
-import { looksLikeScopus } from "../../../utils/scopus";
-import FileErrorMap from "../errors";
+import useError from "../../hooks/useError";
+import useUpload from "../../hooks/useUpload";
+import { looksLikeIsi } from "../../utils/isi";
+import { looksLikeScopus } from "../../utils/scopus";
+import FileErrorMap from "./errors";
 import { FC, useCallback } from "react";
 import { useDropzone } from "react-dropzone";
 import { useLocation, useNavigate } from "react-router-dom";
