+++ src/components/upload/FileErrors.tsx
@@ -1,6 +1,6 @@
-import FileContext from "../../../context/FileContext";
-import useInvalidFiles from "../../../hooks/useInvalidFiles";
-import CancelIcon from "../../vectors/CancelIcon";
+import FileContext from "../../context/FileContext";
+import useInvalidFiles from "../../hooks/useInvalidFiles";
+import CancelIcon from "../vectors/CancelIcon";
 import { FC, useContext } from "react";
 
 const FileErrors: FC = () => {
